Adding Pylint to travis checks
This commit adds a Pylint check job to travis.  A few files are
excluded because they either cause Pylint to blow up or, in the
case of jirasync it is finding a false positive that we don't
want to exclude as a test elsewhere.  A few noisy false positive
errors are also excluded.

Also remove autoescape as jinja templates are escaped elsewhere.